TER-100 make target to go mod tidy workspace
when using workspace, I found that it has become a bit more difficult
to run go mod tidy as we have to do that on each of the workspace entity
and then run go work sync on root.
so to simplify that, this change introduces make target to run these
sequence of commands.

So now to run go mod tidy in entire workspace, one can do:

```sh
make mod-tidy
make mod-clean mod-tidy # alternatively, to pre-clean .sum files
```