TER-123 parse platform tempalte and generate metadata (#30)
This change implements parser for the Terrarium platform HCL template and the `terrarium.yaml` metadata.

The way this platform parser works is:
- It parses the terraform code using the `github.com/cldcvr/terraform-config-inspect/tfconfig` package
- It traverses the terraform blocks to create a dependency graph starting from components
- It parses the component inputs and output metadata

This change also integrates the updated parser into the `terrarium platform lint` CLI command. Such that, now the lint command checks for errors and also updates the `terrarium.yaml` metadata on success.