TER-236 set module namespace from config
this fixes issue an issue where namespace was not getting updated
it was because the db migration was not adding the default namespace
value to the schema.

now, instead of relying on the db column default, we set the default
namespace as farm repo path and initialize it in the config so that in
future, this can be set to a custom value for an org.