TER-241 Allow reusing workdir in harvest
Add optional --workdir (-w) flag to harvest command.
If set, the harvested modules will be placed to this directory
instead of a temporary location.
The advantage of using the workdir is that the initialized
modules stay together with installed plugins making
subsequent runs of other harvest commands significantly faster
as the plugins do not need to be installed in 'terraform init'.

By default this is set to "" falling back to a temporary location
unique to every run.

Changed the Makefile to use workdir by default set to /tmp/harvest.