TER-283 Fix cross-platform release build for CGO (#94)
The root issue here is that once we switche to CGO_ENABLED=1, the
cross compile across OS/Arch became significantly more complex.
In order to accomplish this, much more build infrastructure has to
be in place - especially on Linux which is what we use for our
build pipelines.

This implements the cross-platform build using a tool called xgo.
This tool was built specifically to solve this problem. I intended
to integrate this more directly into the project Make setup but this
was proving to add even more complexity. As such, it is run from a
bash script - create-release-artifact.sh. The script uses xgo to
create the relevant binaries and packages them up as individual tar.gz
(or zip).

This also includes adjustments to the Github action release.yml to
call the script to generate the release artifacts, create a Github
release and then upload all the artifacts. Note: this is change to
the current action where the release is created manually and the
pipeline is triggered. Here a tag is pushed which kicks off the
release process.

Given these changes, the cross-compile builds in the existing
Makefile have been removed as they aren't relevant/functional
anymore.