TER-299 Add SHA256 calculation to release process (#106)
We will need the SHAs for each release artifact to leverage tools
like Homebrew for installation. This adds the calculation of each
SHA to the process and pushes the resulting file up with the
release.

Also noticed that the syntax in the pre-commit.yaml of trufflehog
was incorrect - that was fixed here as it was blocking commits.