TER-361 remove nil taxonomy pre-insert
the nil taxonomy row that we pre-insert in the migration code
is not needed and is causing issues in restoring the farm db.
so this change removes the pre-insert
also, for debugging convenience, this change adds a hidden flag on
farm update command to specify a local dump file instead of
pulling from the github release.