TER-410 fix edge t8 generate edge case
there was an edge case where the t8 generate command was failing for a
specific platform.

the case use to occur specifically when there is a terraform block
that is not linked to a file. and the platform path is an absolute path
not relative path.

this change fixes the issue by skipping such terraform blocks.