TER-412 regenerate platform metadata on lint
change the behavior of lint command to always regenerate the
platform metadata file (terrarium.yaml) on every lint command execution

This implies that the manual changes in platform metadata file are not
acceptable now.
this also impacts the behavior of terrarium generate command such that
it doesn't respect the manual changes in the platform metadata file

This also impacts the example platform since it was not
actually implementing the dependencies and it was only
added to the metadata file to bypass the errors.
Therefore this change also has updates the the `example/platform`
such that now it implements each dependency in terraform and doesn't
have manually adjusted metadata.
Although the `example/platform` contains working mock implementation
such that it can be used to test the t8 lint & generate command and can
also run terraform plan and apply successfully but it'll generate
random mock values for output instead of actually provisioning resources
on cloud.